# 3.0.3, faster simd path fill