… tests/contexts/draw_image_rhino.png
… tests/contexts/draw_image_rhino2.png