before:
`svg decode ........................ 17.021 ms     17.161 ms    ±0.111   x291`

after:
`svg decode ........................ 16.297 ms     16.472 ms    ±0.104   x304`